UI: added separate module for popups
